@@ -1,56 +1,86 @@
 {
-    "name":"contao-bootstrap/navbar",
-    "description":"Bootstrap navbar component for Contao CMS",
-    "keywords":["bootstrap", "components", "contao", "navbar"],
-    "type":"contao-bundle",
-    "license":"LGPL-3.0-or-later",
-    "authors":[
-        {
-            "name":"David Molineus",
-            "email":"mail@netzmacht.de",
-            "homepage":"https://netzmacht.de",
-            "role":"Developer"
-        }
-    ],
-    "support":{
-        "email":"mail@netzmacht.de",
-        "issues":"https://github.com/contao-bootstrap/navbar/issues",
-        "wiki":"http://contao-bootstrap.de",
-        "source":"https://github.com/contao-bootstrap/navbar"
+  "name": "contao-bootstrap/navbar",
+  "description": "Bootstrap navbar component for Contao CMS",
+  "license": "LGPL-3.0-or-later",
+  "type": "contao-bundle",
+  "keywords": [
+    "bootstrap",
+    "components",
+    "contao",
+    "navbar"
+  ],
+  "authors": [
+    {
+      "name": "David Molineus",
+      "email": "mail@netzmacht.de",
+      "homepage": "https://netzmacht.de",
+      "role": "Developer"
     },
-    "require":{
-        "php":">=7.1",
-        "contao/core-bundle":"^4.4",
-        "contao-bootstrap/core": "^2.0",
-        "menatwork/contao-multicolumnwizard":"^3.3.4",
-        "contao-community-alliance/meta-palettes":"^2.0 || ^1.11",
-        "netzmacht/html": "~2.0",
-        "symfony/http-kernel": "^3.3 || ^4.0"
-    },
-    "require-dev": {
-        "contao/manager-plugin": "^2.0",
-        "phpcq/all-tasks": "^1.2",
-        "php-http/guzzle6-adapter": "^1.1"
-    },
-    "conflict": {
-        "netzmacht/contao-bootstrap": "*"
-    },
-    "autoload": {
-        "psr-4": { "ContaoBootstrap\\Navbar\\": "src/" }
-    },
-    "extra":{
-        "contao": {
-            "transifex": {
-                "project": "contao-bootstrap",
-                "prefix": "navbar-",
-                "languages_cto": "src/Resources/contao/languages",
-                "languages_tx": ".tx"
-            }
-        },
-        "branch-alias": {
-            "dev-master": "2.1.x-dev",
-            "dev-develop": "2.2.x-dev"
-        },
-        "contao-manager-plugin": "ContaoBootstrap\\Navbar\\ContaoManager\\Plugin"
+    {
+      "name": "Alexander Kehr",
+      "email": "info@kehr-solutions.de",
+      "homepage": "https://www.kehr-solutions.de",
+      "role": "Developer"
     }
+  ],
+  "support": {
+    "email": "mail@netzmacht.de",
+    "issues": "https://github.com/contao-bootstrap/navbar/issues",
+    "wiki": "http://contao-bootstrap.de",
+    "source": "https://github.com/contao-bootstrap/navbar"
+  },
+  "require": {
+    "php": "^7.4 || ^8.0",
+    "contao-bootstrap/core": "^2.0",
+    "contao-community-alliance/meta-palettes": "^2.0",
+    "contao/core-bundle": "^4.9",
+    "menatwork/contao-multicolumnwizard-bundle": "^3.4",
+    "netzmacht/contao-toolkit": "^3.7",
+    "netzmacht/html": "^2.0.2",
+    "symfony/config": "^4.4 || ^5.4",
+    "symfony/dependency-injection": "^4.4 || ^5.4",
+    "symfony/http-foundation": "^4.4 || ^5.4",
+    "symfony/http-kernel": "^4.4 || ^5.4",
+    "symfony/routing": "^4.4 || ^5.4",
+    "symfony/translation-contracts": "^1.1 || ^2.0"
+  },
+  "require-dev": {
+    "contao/manager-plugin": "^2.0",
+    "doctrine/coding-standard": "^9.0.",
+    "phpcq/coding-standard": "^2.1.2",
+    "phpcq/plugin-api": "1.x-dev",
+    "phpcq/runner-bootstrap": "1.x-dev"
+  },
+  "conflict": {
+    "netzmacht/contao-bootstrap": "*"
+  },
+  "autoload": {
+    "psr-4": {
+      "ContaoBootstrap\\Navbar\\": "src/"
+    }
+  },
+  "config": {
+    "allow-plugins": {
+      "contao-components/installer": false,
+      "contao/manager-plugin": true,
+      "dealerdirect/phpcodesniffer-composer-installer": true,
+      "contao-community-alliance/composer-plugin": true
+    },
+    "sort-packages": true
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-develop": "2.3.x-dev",
+      "dev-master": "2.2.x-dev"
+    },
+    "contao": {
+      "transifex": {
+        "languages_cto": "src/Resources/contao/languages",
+        "languages_tx": ".tx",
+        "prefix": "navbar-",
+        "project": "contao-bootstrap"
+      }
+    },
+    "contao-manager-plugin": "ContaoBootstrap\\Navbar\\ContaoManager\\Plugin"
+  }
 }
