@@ -1,12 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 /**
  * Contao Bootstrap Navbar.
  *
- * @package    contao-bootstrap
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2017-2018 netzmacht David Molineus. All rights reserved.
- * @license    LGPL 3.0-or-later
  * @filesource
  */
 
@@ -16,6 +14,9 @@ use Contao\StringUtil;
 use Netzmacht\Html\Attributes;
 use Netzmacht\Html\Exception\InvalidArgumentException;
 
+use function implode;
+use function in_array;
+
 /**
  * Class HeaderItemHelper creates the header navigation item.
  */
@@ -24,21 +25,19 @@ final class HeaderItemHelper extends Attributes implements ItemHelper
     /**
      * Current item.
      *
-     * @var array
+     * @var array<string,mixed>
      */
-    protected $item;
+    protected array $item;
 
     /**
      * Item classes.
      *
-     * @var array
+     * @var list<string>
      */
-    protected $itemClass = array();
+    protected array $itemClass = [];
 
     /**
-     * AbstractItemHelper constructor.
-     *
-     * @param array $item Navigation item.
+     * @param array<string,mixed> $item Navigation item.
      *
      * @throws InvalidArgumentException When invalid attributes are given.
      */
@@ -64,9 +63,6 @@ final class HeaderItemHelper extends Attributes implements ItemHelper
         return implode(' ', $this->itemClass);
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function getTag(): string
     {
         return 'div';
@@ -74,20 +70,22 @@ final class HeaderItemHelper extends Attributes implements ItemHelper
 
     /**
      * Initialize the item classes.
-     *
-     * @return void
      */
     private function initializeItemClasses(): void
     {
-        if ($this->item['class']) {
-            $classes = StringUtil::trimsplit(' ', $this->item['class']);
-            foreach ($classes as $class) {
-                $this->itemClass[] = $class;
-            }
-
-            if (in_array('trail', $this->itemClass)) {
-                $this->itemClass[] = 'active';
-            }
+        if (! $this->item['class']) {
+            return;
         }
+
+        $classes = StringUtil::trimsplit(' ', $this->item['class']);
+        foreach ($classes as $class) {
+            $this->itemClass[] = $class;
+        }
+
+        if (! in_array('trail', $this->itemClass)) {
+            return;
+        }
+
+        $this->itemClass[] = 'active';
     }
 }
