@@ -1,15 +1,5 @@
 <?php
 
-/**
- * Contao Bootstrap Navbar.
- *
- * @package    contao-bootstrap
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2017 netzmacht David Molineus. All rights reserved.
- * @license    LGPL 3.0
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Navbar\ContaoManager;
@@ -23,10 +13,8 @@ use ContaoBootstrap\Navbar\ContaoBootstrapNavbarBundle;
 
 /**
  * Contao manager plugin.
- *
- * @package ContaoBootstrap\Navbar\ContaoManager
  */
-class Plugin implements BundlePluginInterface
+final class Plugin implements BundlePluginInterface
 {
     /**
      * {@inheritdoc}
