@@ -1,15 +1,5 @@
 <?php
 
-/**
- * Contao Bootstrap Navbar.
- *
- * @package    contao-bootstrap
- * @author     David Molineus <david.molineus@netzmacht.de>
- * @copyright  2017 netzmacht David Molineus. All rights reserved.
- * @license    LGPL 3.0
- * @filesource
- */
-
 declare(strict_types=1);
 
 namespace ContaoBootstrap\Navbar\Helper;
@@ -17,7 +7,7 @@ namespace ContaoBootstrap\Navbar\Helper;
 /**
  * Navigation item helper for a nav item.
  *
- * @package ContaoBootstrap\Navbar\Helper
+ * @psalm-suppress PropertyNotSetInConstructor - False detected issues. Parent constructors initializes them
  */
 class NavItemHelper extends AbstractItemHelper
 {
@@ -31,13 +21,15 @@ class NavItemHelper extends AbstractItemHelper
         $this->addClass('nav-link');
         $this->itemClass[] = 'nav-item';
 
-        if ($this->item['subitems']) {
-            $this->itemClass[] = 'dropdown';
-            $this->addClass('dropdown-toggle');
-
-            $this->setAttribute('data-toggle', 'dropdown');
-            $this->setAttribute('aria-haspopup', 'true');
-            $this->setAttribute('aria-expanded', 'false');
+        if (! $this->item['subitems']) {
+            return;
         }
+
+        $this->itemClass[] = 'dropdown';
+        $this->addClass('dropdown-toggle');
+
+        $this->setAttribute('data-toggle', 'dropdown');
+        $this->setAttribute('aria-haspopup', 'true');
+        $this->setAttribute('aria-expanded', 'false');
     }
 }
